+++ examples/tanstack-example/vite.config.ts
@@ -1,10 +1,13 @@
-import { tanstackStart } from '@tanstack/react-start/plugin/vite'
-import { defineConfig } from 'vite'
-import viteReact from '@vitejs/plugin-react'
-import tailwindcss from '@tailwindcss/vite'
-import { nitro } from 'nitro/vite'
+import { tanstackStart } from "@tanstack/react-start/plugin/vite";
+import { defineConfig } from "vite";
+import viteReact from "@vitejs/plugin-react";
+import tailwindcss from "@tailwindcss/vite";
+import { nitro } from "nitro/vite";
 
 export default defineConfig({
+  optimizeDeps: {
+    include: ["tsdkarc", "tsdkarc-x"],
+  },
   server: {
     port: 3000,
   },
@@ -14,9 +17,9 @@ export default defineConfig({
   plugins: [
     tailwindcss(),
     tanstackStart({
-      srcDirectory: 'src',
+      srcDirectory: "src",
     }),
     viteReact(),
     nitro(),
   ],
-})
+});