+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # TsdkArc Packages Changelog
 
+
+## tsdkarc-x@0.1.2, tsdkarc@2.0.0-alpha.2, tsdkbundle@0.0.2
+
+- Fix esm bundle output(need `.js` ext for vite)
+
+- Fix `tsdkarc-x/esm` not build
+
 ## tsdkarc-x@0.1.1
 
 - Fix `tsdkarc-x/esm` not build